feat: renamed bondAPR to argonBondAPR for clarity + bitcoinAPR
